requirements.txt
aiohttp
apscheduler
cinemagoer
cloudscraper
colour
cowpy
emoji
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gtts
hachoir
heroku3
html-telegraph-poster
httpx[http2]
humanize
jikanpy
justwatch
lottie
lyricsgenius
markdown
motor
openai
Pillow
prettytable
psutil
psycopg2
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
pytz
qrcode
requests
selenium
setuptools
ShazamAPI
somnium
spamwatch
speedtest-cli
telegraph
tgcrypto
ujson
urlextract
validators
vcsi
wand
wget
python-dotenv

git+https://github.com/Jisan09/search-engine-parser
git+https://github.com/LonamiWebs/Telethon.git
git+https://github.com/goldsmith/Wikipedia
git+https://github.com/sandy1709/py-googletrans
git+https://github.com/alexmercerind/youtube-search-python
git+https://github.com/yt-dlp/yt-dlp
